Refactor and restructure.
Publishing is now fully multi-threaded. Publishing is now handled through a
publisher class, which can be used to customize the behavior (for instance,
not publishing certain node kinds.) The template repository is now pickable,
which is a pre-requisite for multiprocessing (wasn't pickable previously, but
as documents were not published in parallel, nobody noticed.)

Various classes have been moved around into separate files to simplify
maintenance.