Move has-pending-document into a plugin.
This introduces Liara.setup_plugins(), a static method to discover
plugins. The command line will call this to give plugins a chance to
register themselves with the command line parser. This has the downside
of executing before the arguments are parsed, so debug output from the
setup will be missing as logging hasn't been set up yet.