Improve site/page queries & error handling.
In templates, it's usually desired to access only index and document
nodes, but so far there was no way to do so. This change adds new
accessors as well as filtering functions to that effect.

Additionally, it improves the failure mode with better error messages
and extended logging (in debug mode.)